+++ srai/utils/_pytorch_stubs.py
@@ -1,3 +1,14 @@
+from typing import Generic, TypeVar
+
+T_co = TypeVar("T_co", covariant=True)
+
+
+class Dataset(Generic[T_co]):
+    """Dataset class stub."""
+
+    ...
+
+
 class LightningModule:  # pragma: no cover
     """LightningModule class stub."""